requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
dbfread==2.0.7
psycopg2-binary==2.9.9
python-multipart==0.0.9
rapidfuzz==3.8.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
